fixes for WSL2 path problems: use a docker volume instead of a bind mount because WSL2 has some path problems with /mnt/c vs. /c and neither of those really seems to work right, so an external volume mounted at /data fixes the file permissions and path problems.
the caveat is that the files tarsnap accesses must be 'docker cp' into and out of the volume.

and yet, docker secrets work normally!

another bug: tarsnap prints many errors of the form   Can't translate pathname x to UTF-8 . I set LANG=C to try to fix it, and the files get copied, but it doesn't suppress the errors due to UTF characters